requirements.txt
# Web Framework
Flask==3.1.0

#Flask Error
Werkzeug==3.1.3

# HTTP Requests
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

